requirements.txt
Flask
PyPDF2
python-pptx
google-generativeai
python-dotenv